+++ iota-bundle-miner/src/lib.rs
@@ -3,9 +3,11 @@
 
 pub mod constant;
 pub mod error;
+pub mod helper;
 pub mod miner;
 pub mod recoverer;
 pub mod success;
 
 pub use miner::MinerBuilder;
+pub use recoverer::RecovererBuilder;
 pub use success::success;
